created interface for each service class
