+++ trading/src/main/java/com/cryptocurrency/trading/Service/AccountResetService.java
@@ -1,20 +1,8 @@
 package com.cryptocurrency.trading.Service;
 
 import jakarta.transaction.Transactional;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
 
-@Service
-@RequiredArgsConstructor
-public class AccountResetService {
-    private final UserService userService;
-    private final UserAssetService userAssetService;
-    private final TransactionService transactionService;
-
+public interface AccountResetService {
     @Transactional
-    public void resetUserAccount(int userId) throws Exception {
-        userService.resetBalance(userId);
-        userAssetService.removeUserAssetsByUserId(userId);
-        transactionService.removeTransactionsByUserId(userId);
-    }
+    void resetUserAccount(int userId) throws Exception;
 }
